+++ src/main/java/act/cli/CliContext.java
@@ -346,7 +346,7 @@ public class CliContext extends ActContext.Base<CliContext> implements IASCIITab
         String label = app().config().i18nEnabled() ? i18n("act.progress.capFirst") : "Progress";
         ProgressBar pb = new ProgressBar(label, progressGauge.maxHint(), 200, os, ProgressBarStyle.UNICODE_BLOCK);
         pb.start();
-        while (!progressGauge.done()) {
+        while (!progressGauge.isDone()) {
             pb.maxHint(progressGauge.maxHint());
             pb.stepTo(progressGauge.currentSteps());
             flush();
@@ -358,7 +358,7 @@ public class CliContext extends ActContext.Base<CliContext> implements IASCIITab
     public void printText(ProgressGauge progressGauge) {
         SimpleProgressGauge simpleProgressGauge = SimpleProgressGauge.wrap(progressGauge);
         boolean i18n = app().config().i18nEnabled();
-        while (!progressGauge.done()) {
+        while (!progressGauge.isDone()) {
             if (i18n) {
                 print("\r" + i18n("act.progress.report", simpleProgressGauge.currentProgressPercent()));
             } else {