+++ src/main/java/act/conf/AppConfigKey.java
@@ -151,11 +151,12 @@ public enum AppConfigKey implements ConfigKey {
     CLI_PAGE_SIZE_TABLE("cli.page.size.table"),
 
     /**
-     * {@code cli.session.expiration} specifies the number of seconds
+     * {@code cli.session.ttl} specifies the number of seconds
      * a cli session can exists after last user interaction
+     *
      * <p>Default value: {@code 300} seconds. e.g. 5 minutes</p>
      */
-    CLI_SESSION_EXPIRATION("cli.session.expiration.int"),
+    CLI_SESSION_TTL("cli.session.ttl.int"),
 
     /**
      * {@code cli.session.max} specifies the maximum number of cli threads